sys/syscall/enter.c: hack support for mounting sysfs
This is a total hack where when we are mounting sysfs somewhere we actually just create a symlink there instead and change the arguments to the syscall to be a fake bind mount - this always works and ensures that the tracee thinks it's /sys is working as expected. There is probably a much more elegant way to handle this, but this hack works for now.